Bug fix

See merge request robo-gym/robo-gym!108

+++ robo_gym/envs/mir100/mir100.py
@@ -119,6 +119,9 @@ class Mir100Env(gym.Env):
         return 0, False, {}
 
     def step(self, action):
+        
+        action = action.astype(np.float32)
+
         self.elapsed_steps += 1
 
         # Check if the action is within the action space
@@ -269,7 +272,7 @@ class Mir100Env(gym.Env):
             # Compose environment state
             state = np.concatenate((np.array([polar_r, polar_theta]),rs_state[6:8]))
 
-        return state
+        return state.astype(np.float32)
 
     def _get_observation_space(self):
         """Get environment observation space.
